+++ src/main/java/com/github/rodionovsasha/commentservice/services/impl/CommentServiceImpl.java
@@ -23,33 +23,33 @@ public class CommentServiceImpl implements CommentService {
     private final CommentRepository repository;
 
     @Override
-    public Comment add(String content, long topicId, long userId) {
+    public Comment add(String content, int topicId, int userId) {
         val user = userService.getActiveUser(userId);
         val topic = topicService.getActiveTopic(topicId);
         return repository.save(new Comment(content, user, topic));
     }
 
     @Override
-    public void update(long commentId, long userId, String content) {
+    public void update(int commentId, int userId, String content) {
         update(commentId, userId, comment -> comment.setContent(content));
     }
 
     @Override
-    public void archive(long commentId, long userId) {
+    public void archive(int commentId, int userId) {
         update(commentId, userId, comment -> comment.setArchived(true));
     }
 
     @Override
-    public List<Comment> findByTopic(long topicId) {
+    public List<Comment> findByTopic(int topicId) {
         topicService.checkTopicExists(topicId);
         return repository.findByTopicIdAndArchivedFalseOrderByDateAsc(topicId);
     }
 
-    private Comment getById(long id) {
+    private Comment getById(int id) {
         return repository.findOne(id).orElseThrow(() -> CommentNotFoundException.forId(id));
     }
 
-    private void update(long commentId, long userId, Consumer<Comment> consumer) {
+    private void update(int commentId, int userId, Consumer<Comment> consumer) {
         userService.checkUserActive(userId);
         val comment = getById(commentId);
         if (comment.getUser().getId() != userId) {
